Finished merging and began abstracting components so styling can be worked on independently of data processing
