+++ components/CollapsibleTable.module.css
@@ -1,3 +1,4 @@
 .tableRow{
-    color: blue
+    color: blue;
+	width: "100vw";
 }
