+++ components/CollapsibleTable.tsx
@@ -15,105 +15,80 @@ import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
 import styles from "./CollapsibleTable.module.css";
 import Tabs from "./Tabs";
 
-function createData(name: string, version: string, link: string) {
-  {
-    return {
-      name,
-      version,
-      link,
-    };
-  }
-}
+import Row from "./Row"
 
-// Sample data for the table
-const rows = [
-  createData(
-    "@octokit/app",
-    "12.0.5",
-    "https://github.com/octokit/app.js/tree/master"
-  ),
-  createData(
-    "@octokit/core",
-    "3.6.0",
-    "https://github.com/octokit/core.js/tree/master"
-  ),
-  createData(
-    "@octokit/oauth-app",
-    "3.6.0",
-    "https://github.com/octokit/oauth-app.js/tree/master"
-  ),
-  createData(
-    "@octokit/plugin-paginate-rest",
-    "2.17.0",
-    "https://github.com/octokit/plugin-paginate-rest.js/tree/master"
-  ),
-];
+// function createData(name: string, version: string, link: string) {
+// 	{
+// 		return {
+// 			name,
+// 			version,
+// 			link,
+// 		};
+// 	}
+// }
 
-// Creates each invidial row
-const Row = (props: { row: ReturnType<typeof createData> }) => {
-  const { row } = props;
-  const [open, setOpen] = useState(false);
+// // Sample data for the table
+// const rows = [
+// 	createData(
+// 		"@octokit/app",
+// 		"12.0.5",
+// 		"https://github.com/octokit/app.js/tree/master"
+// 	),
+// 	createData(
+// 		"@octokit/core",
+// 		"3.6.0",
+// 		"https://github.com/octokit/core.js/tree/master"
+// 	),
+// 	createData(
+// 		"@octokit/oauth-app",
+// 		"3.6.0",
+// 		"https://github.com/octokit/oauth-app.js/tree/master"
+// 	),
+// 	createData(
+// 		"@octokit/plugin-paginate-rest",
+// 		"2.17.0",
+// 		"https://github.com/octokit/plugin-paginate-rest.js/tree/master"
+// 	),
+// ];
 
-  return (
-    <div>
-      <TableRow className={styles.tableRow}>
-        <TableCell>
-          <IconButton
-            aria-label="expand row"
-            size="small"
-            onClick={() => setOpen(!open)}
-          >
-            {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
-          </IconButton>
-        </TableCell>
-        <TableCell component="th" scope="row">
-          {row.name}
-        </TableCell>
-        <TableCell align="right">{row.version}</TableCell>
-        <TableCell align="right">
-          <a href={row.link} target="_blank">
-            Open In GitHub
-          </a>
-        </TableCell>
-      </TableRow>
-      <TableRow>
-        <TableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={6}>
-          <Collapse in={open} timeout="auto" unmountOnExit>
-            <Box sx={{ margin: 1 }}>
-              <Table size="small" aria-label="purchases">
-                <TableHead>
-                  <TableRow>
-                    <TableCell>
-                      <Tabs></Tabs>
-                    </TableCell>
-                  </TableRow>
-                </TableHead>
-              </Table>
-            </Box>
-          </Collapse>
-        </TableCell>
-      </TableRow>
-    </div>
-  );
-};
+// // Creates the whole table based on sample data 
+// const CollapsibleTable = () => {
+// 	const x = rows.map((row) => (<Row key={row.name} row={row} />))
+// 	return makeCollapsibleTable(x)
+// };
 
 // Creates the whole table based on sample data 
-const CollapsibleTable = () => {
-  return (
-    <div>
-      <TableContainer component={Paper}>
-        <Table aria-label="collapsible table">
-          <TableHead>
-          </TableHead>
-          <TableBody>
-            {rows.map((row) => (
-              <Row key={row.name} row={row} />
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    </div>
-  );
+const CollapsibleTable = (rows: JSX.Element[]) => {
+	return (
+		<div>
+			<TableContainer component={Paper}>
+				<Table size="small" aria-label="collapsible table">
+					<colgroup>
+						<col style={{width:'0%', backgroundColor: "#f6f6f6"}}/>
+						<col style={{width:'75%'}}/>
+						<col style={{width:'25%'}}/>
+						<col style={{width:'0%'}}/>
+					</colgroup>
+					<TableHead>
+						<TableRow>
+							<TableCell></TableCell>
+							<TableCell>Name</TableCell>
+							<TableCell>Version</TableCell>
+							<TableCell></TableCell>
+						</TableRow>
+					</TableHead>
+					<TableBody>
+						{rows}
+					</TableBody>
+				</Table>
+			</TableContainer>
+		</div>
+	);
 };
 
-export default CollapsibleTable;
+const makeCollapsibleTable = (JSObject: [Map<number, any>, any] ) => {
+	//TODO: Rather than taking in the raw objet, take a structure that can be directly converted into JSX
+	return CollapsibleTable( Object.entries(JSObject[0]).map(([id, row]) => (<Row key={row.name} row={row} />)));
+};
+
+export default makeCollapsibleTable;
