+++ components/Colours.ts
@@ -1,0 +1,11 @@
+export const Colours = {
+	green: "#cfc" as any,
+	greenBorder: "1px solid green" as any,
+	orange: "#ffc" as any,
+	orangeBorder: "1px solid orange" as any,
+	red: "#fcc" as any,
+	redBorder: "1px solid red" as any,
+	ahmThinBar: "#ccc" as any,
+	ahmLeftTable: "#f6f6f6" as any,
+	ahmBlack: "#000" as any,
+}
