+++ components/Row.tsx
@@ -1,0 +1,66 @@
+import React, { useState } from "react";
+import Box from "@mui/material/Box";
+import Collapse from "@mui/material/Collapse";
+import IconButton from "@mui/material/IconButton";
+import Table from "@mui/material/Table";
+import TableBody from "@mui/material/TableBody";
+import TableCell from "@mui/material/TableCell";
+import TableContainer from "@mui/material/TableContainer";
+import TableHead from "@mui/material/TableHead";
+import TableRow from "@mui/material/TableRow";
+import Paper from "@mui/material/Paper";
+import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
+import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
+
+import styles from "./CollapsibleTable.module.css";
+import Tabs from "./Tabs";
+
+
+// Creates each invidial row
+const Row = (props: { row: { name: string, version: string, link: string } }) => {
+	const { row } = props;
+	const [open, setOpen] = useState(false);
+
+	return (
+		<React.Fragment>
+			<TableRow className={styles.tableRow}>
+				<TableCell>
+					<IconButton
+						aria-label="expand row"
+						size="small"
+						onClick={() => setOpen(!open)}
+					>
+						{open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
+					</IconButton>
+				</TableCell>
+				<TableCell component="th" scope="row">{row.name}</TableCell>
+				<TableCell align="right">{row.version}</TableCell>
+				<TableCell align="right">
+					(
+					<a href={row.link} target="_blank">
+						GitHub
+					</a>
+					)
+				</TableCell>
+			</TableRow>
+			<TableRow>
+				<TableCell style={{ paddingBottom: 0, paddingTop: 0, backgroundColor: "#FFF" }} colSpan={6}>
+					<Collapse in={open} timeout="auto" unmountOnExit>
+						<Box sx={{ margin: 1 }}>
+							<Table size="small" aria-label="purchases">
+								<TableHead>
+									<TableRow>
+										<TableCell>
+											<Tabs></Tabs>
+										</TableCell>
+									</TableRow>
+								</TableHead>
+							</Table>
+						</Box>
+					</Collapse>
+				</TableCell>
+			</TableRow>
+		</React.Fragment>
+	);
+};
+export default Row;
