+++ pages/home.tsx
@@ -1,10 +1,20 @@
 import React from 'react'
-import CollapsibleTable from '../components/CollapsibleTable'
+import makeCollapsibleTable from '../components/CollapsibleTable'
+import Row from '../components/Row'
 
 const home = () => {
-  return (
-    <div><CollapsibleTable/></div>
-  )
+	return (
+		<div>
+			{
+				// CollapsibleTable(
+				// 	[
+				// 		<Row key={"a"} row={{ name: "1", version: "2", link: "3" }} />,
+				// 		<Row key={"B"} row={{ name: "4", version: "5", link: "6" }} />
+				// 	]
+				// )
+			}
+		</div>
+	)
 }
 
 export default home
